modify quality type CGFloat to flaot

+++ TCacheServiceDemo/TCacheService/TImageCacheService.h
@@ -19,7 +19,7 @@
  */
 - (void)cacheData:(id)data
               key:(NSString *)key
-          quality:(CGFloat)quality;
+          quality:(float)quality;
 
 /**
  缓存图片
@@ -32,7 +32,7 @@
 - (void)cacheData:(id)data
               key:(NSString *)key
          extrakey:(NSString*)extraKey
-          quality:(CGFloat)quality;
+          quality:(float)quality;
 
 /**
  异步缓存
@@ -44,13 +44,13 @@
  */
 - (void)cacheData:(id)data
               key:(NSString *)key
-          quality:(CGFloat)quality
+          quality:(float)quality
          complete:(dispatch_block_t)complete;
 
 - (void)cacheData:(id)data
               key:(NSString *)key
          extraKey:(NSString*)extraKey
-          quality:(CGFloat)quality
+          quality:(float)quality
          complete:(dispatch_block_t)complete;
 
 @end
